Rifle class - getAngle() returns float

+++ core/src/com/rifle/game/Rifle.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.math.Vector2;
 public class Rifle {
 	private Vector2 position;
 	private World world;
-	private double angle;
+	private float angle;
 	 
     public Rifle(int x, int y) {
         position = new Vector2(x,y);
@@ -19,13 +19,13 @@ public class Rifle {
     public void updateAngle() {
     	Vector2 cursorPos = new Vector2(Gdx.input.getX(), RifleGame.HEIGHT - Gdx.input.getY());
     	Vector2 rifleCenterPos = position;
-    	angle = Math.atan2(cursorPos.y - rifleCenterPos.y, cursorPos.x - rifleCenterPos.x) * (180/Math.PI);
+    	angle = (float) (Math.atan2(cursorPos.y - rifleCenterPos.y, cursorPos.x - rifleCenterPos.x) * (180/Math.PI));
     	if (angle < 0) {
     		angle += 360;
     	}
     }
     
-    public double getAngle() {
+    public float getAngle() {
     	updateAngle();
     	return angle;
     }
